remote queries: raise multipart forwarding caps to 10 GiB and independent part size
Expand the Agent multipart result-delivery caps to the Phase 2 targets and
decouple COPY read size from multipart part size:

- multipart part hard cap 64 MiB -> 128 MiB;
- multipart total hard cap 1 GiB -> 10 GiB;
- remove the invalid result_delivery.partBytes <= copyLimits.chunkBytes
  validation: COPY reads and multipart parts are independent, and multiple
  bounded COPY reads aggregate into one multipart part;
- retain result_delivery.maxBytes <= copyLimits.maxBytes as the overall
  extraction safety ceiling only.

10 GiB overflows int32, so RemoteQueryExecuteCopyLimits.max_bytes and
RemoteQueryResultDelivery.max_bytes are widened to int64; part_bytes and
chunk_bytes stay int32 since the 128 MiB part cap and 1 MiB COPY read fit
comfortably. RemoteQueryExecuteLimits.max_bytes (inline path) is unchanged.

The Agent remains a control-plane forwarder: it still carries baseUrl and the
scoped token through to the integration request JSON and surfaces only the
compact receipt, so bulk bytes never traverse AgentSecure/PAR/AP. Phase 1
field names and forwarding/auth/redaction behavior are preserved.

Adds boundary tests for the exact 10 GiB and 128 MiB caps and cap-plus-one
rejections, plus JSON/proto int64 fidelity tests proving 10 GiB round-trips
through the AgentSecure proto and the integration request JSON without loss,
and that a 10 GiB upload is accepted when the extraction cap is 10 GiB while an
upload limit exceeding the extraction cap is rejected.

Regenerates pkg/proto/pbgo/core/api.pb.go from api.proto.

(cherry picked from commit cbbc7c898365314216bf046a96bbc31d372d12b9)

+++ pkg/privateactionrunner/bundles/remoteaction/queries/execute.go
@@ -221,7 +221,7 @@ func remoteQueryExecuteRequestFromInputs(inputs ExecuteInputs) *pb.RemoteQueryEx
 	if inputs.CopyLimits != nil {
 		req.CopyLimits = &pb.RemoteQueryExecuteCopyLimits{
 			ChunkBytes:  int32(inputs.CopyLimits.ChunkBytes),
-			MaxBytes:    int32(inputs.CopyLimits.MaxBytes),
+			MaxBytes:    int64(inputs.CopyLimits.MaxBytes),
 			MaxRowBytes: int32(inputs.CopyLimits.MaxRowBytes),
 			TimeoutMs:   int32(inputs.CopyLimits.TimeoutMs),
 		}
@@ -233,7 +233,7 @@ func remoteQueryExecuteRequestFromInputs(inputs ExecuteInputs) *pb.RemoteQueryEx
 			BaseUrl:     inputs.ResultDelivery.BaseURL,
 			Token:       inputs.ResultDelivery.Token,
 			PartBytes:   int32(inputs.ResultDelivery.PartBytes),
-			MaxBytes:    int32(inputs.ResultDelivery.MaxBytes),
+			MaxBytes:    int64(inputs.ResultDelivery.MaxBytes),
 			Format:      inputs.ResultDelivery.Format,
 			Compression: inputs.ResultDelivery.Compression,
 		}
